+++ lib/ZbootaClient.php
@@ -137,4 +137,11 @@ function newUser() {
 	$this->initiateAccount();
 }
 
+function deleteTestUser() {
+	$this->client->deleteItem(array(
+	    'TableName' => 'zboota-users',
+	    'Key' => array( 'email'      => array('S' => "shadiakiki1986@hotmail.com") )
+	));
+}
+
 }